Release v3.1.10

+++ netbox/dcim/models/racks.py
@@ -412,7 +412,7 @@ class Rack(PrimaryModel):
                 available_units.remove(u)
 
         occupied_unit_count = self.u_height - len(available_units)
-        percentage = int(float(occupied_unit_count) / self.u_height * 100)
+        percentage = float(occupied_unit_count) / self.u_height * 100
 
         return percentage
